Move quantum program sources to their final layout
The node contract and the node catalogue now share one module door, so a node
type is reached as `nodes::math::Add` rather than being split between a
top-level `program_node` and a sibling `math_nodes` directory. `tensor` and
`nodes` are directories with a `mod.rs`, matching the rest of `crates/`.

Only module and import declarations change.
